Create cs1.txt
Recode the roster example we studied in class such that it prints the list of classes each student is enrolled in. Here is an example printout:
All Students
last name, first name:  courses enrolled 
Kathleen Anderson: CS1   CS3  CS4
Gerald Edwards: CS1   CS2 CS3
Mary Price: CS1
...
Here is the output for the above example roster files.
The project should be done using associative containers and class Student to hold the name of the student. Specifically, you need to use a map keyed by the class Student while value could be a list of classes the student is enrolled in. The map should be the ordered map. That is, the data structure to be used in as follows: map<Student, list<string>>

Recode the roster example so that it prints all the students currently enrolled in at least one course. That is, enrolled in a course and are not in dropout list. The name of each enrolled student has to be printed exactly once. You have to use this data structure: set<Student>. Refer to this set example for guidance. Here is an example printout:
Currently Enrolled Students
Kathleen Anderson
Gerald Edwards
Mary Price
...